OOP - complete. MyPoint coordinate int -> double.

+++ homework/task1_OOP/evclidGeometry/MyPoint.java
@@ -6,38 +6,38 @@ package education.homework.task1_OOP.evclidGeometry;
  */
 
 public class MyPoint {
-    private int x; //=0
-    private int y; //=0
+    private double x; //=0
+    private double y; //=0
 
     public MyPoint() {
     }
 
-    public MyPoint(int x, int y) {
+    public MyPoint(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public int[] getXY() {
-        return new int[]{x, y};
+    public double[] getXY() {
+        return new double[]{x, y};
     }
 
-    public void setXY(int x, int y) {
+    public void setXY(double x, double y) {
         this.x = x;
         this.y = y;
     }
@@ -48,7 +48,7 @@ public class MyPoint {
                 ',' + y + ')';
     }
 
-    public double distance(int x, int y) {
+    public double distance(double x, double y) {
         return Math.sqrt(Math.pow(this.x - x, 2) + Math.pow(this.y - y, 2));
     } //Distance from this point the given point at (x,y)
